Folders reorganization and Home-Page view settings
