+++ src/js/component/Menus/ColleaguesSideMenu/ClassesComponent/Classes.jsx
