+++ src/js/component/Menus/ColleaguesSideMenu/ClassesComponent/ClassesStyles.scss
