+++ src/js/component/Menus/Resources/Resource.jsx
