+++ src/js/component/Menus/Resources/Resource.scss
