+++ src/js/component/Menus/Resources/ResourcesMenu.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import Resource from "./Resource.jsx";
-import Collapse from "../Buttons/Collapse/Collapse.jsx";
+import Collapse from "../../Buttons/Collapse/Collapse.jsx";
 
 import "./ResourcesMenu.scss";
