+++ src/js/component/Menus/Resources/ResourcesMenu.scss
